requirements.txt
# Core dependencies
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2

# LightGBM
lightgbm==4.1.0

# Bayesian Optimization
bayesian-optimization==1.4.0

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0

# Visualization
streamlit==1.28.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Data processing
scipy==1.11.4

# Logging & Config
pyyaml==6.0.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Docker
supervisor==4.2.5

# Development
black==23.11.0
flake8==6.1.0
jupyter==1.0.0
